requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2.20
django-celery-beat==2.5.0
django-timezone-field==7.1
kombu==5.5.2
prompt_toolkit==3.0.50
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
